+++ pkg/cmd/tasks/tasks.go
@@ -2,6 +2,7 @@ package tasks
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/timwehrle/asana/pkg/cmd/tasks/comments"
 	"github.com/timwehrle/asana/pkg/cmd/tasks/create"
 	"github.com/timwehrle/asana/pkg/cmd/tasks/delete"
 	"github.com/timwehrle/asana/pkg/cmd/tasks/list"
@@ -22,6 +23,7 @@ func NewCmdTasks(f factory.Factory) *cobra.Command {
 
 	cmd.AddCommand(list.NewCmdList(f, nil))
 	cmd.AddCommand(view.NewCmdView(f, nil))
+	cmd.AddCommand(comments.NewCmdComments(f, nil))
 	cmd.AddCommand(update.NewCmdUpdate(f, nil))
 	cmd.AddCommand(search.NewCmdSearch(f, nil))
 	cmd.AddCommand(create.NewCmdCreate(f, nil))